post page no longer needs a logged in user
