+++ views/layouts/main.handlebars
@@ -12,23 +12,43 @@
   <link type="text/css" rel="stylesheet" href="/assets/css/style.css" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Home Page</title>
-  <script type="text/javascript" src="/materialize/js/materialize.min.js"></script>
   <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
   <script src="https://cdn.auth0.com/js/auth0/9.10/auth0.min.js"></script>
-  <script src="/js/app.js"></script>
 </head>
 
 <body>
+  <div id="dropdown1" class="dropdown-content">
+    <div class="row">
+      <div class="input-field col s11">
+        <i class="material-icons prefix black-text">account_circle</i>
+        <input id="username" type="text" class="validate">
+        <label for="username">Username</label>
+      </div>
+      <div class="input-field col s11">
+        <i class="material-icons prefix black-text">security</i>
+        <input id="password" type="password" class="validate">
+        <label for="password">Password</label>
+      </div>
+      <div class="col s11 black-text valign-wrapper">
+        <a class="waves-effect waves-light btn right" id="login-button">Login</a>
+        <span class="right top">OR</span>
+        <a class="waves-effect waves-light btn right" href="/register">Register</a>
+      </div>
+    </div>
+  </div>
   <nav>
     <div class="container wider-container">
-    <div class="nav-wrapper">
-      <a href="#" class="brand-logo">Kids these days</a>
-      <ul id="nav-mobile" class="right hide-on-med-and-down">
-        <li><a href="/posts">Posts</a></li>
-        <li><a href="/profile">Profile</a></li>
-        <li><a href="/api/sessions_logout">Logout</a></li>
-      </ul>
-    </div>
+      <div class="nav-wrapper">
+        <a href="#" class="brand-logo">Kids these days</a>
+        <ul id="nav-mobile" class="right hide-on-med-and-down">
+          {{#if user}}
+          <li><a href="/api/sessions_logout">Logout</a></li>
+          {{else}}
+          <li><a class="dropdown-trigger" href="#!" data-target="dropdown1">Login</a></li>
+          {{/if}}
+          </li>
+        </ul>
+      </div>
     </div>
   </nav>
   <div class="container wider-container">
@@ -60,6 +80,7 @@
       </div>
     </div>
   </footer>
+  <script type="text/javascript" src="/materialize/js/materialize.js"></script>
 </body>
 
 </html>
