+++ views/posts.handlebars
@@ -1,3 +1,9 @@
+{{#if user}}
+<h1>{{user.firstname}} {{user.lastname}}</h1>
+<h1>{{user.email}}</h1>
+{{/if}}
+<br>
+<br>
 <div id="home">
   <div class="row">
     <div class="col s4">
